Rename hydrate & hydrate-support to "experimental"
