+++ packages/lit-element/rollup.config.js
@@ -11,7 +11,7 @@ export default litProdConfig({
   entryPoints: [
     'index',
     'lit-element',
-    'hydrate-support',
+    'experimental-hydrate-support',
     'private-ssr-support',
     'decorators',
     'decorators/custom-element',
