+++ packages/lit/rollup.config.js
@@ -36,8 +36,8 @@ export default litProdConfig({
     'directives/until',
     'async-directive',
     'html',
-    'hydrate-support',
-    'hydrate',
+    'experimental-hydrate-support',
+    'experimental-hydrate',
     'index',
     'static-html',
   ],
